Фикс потери маммонов в Peddler

+++ code/modules/roguetown/roguemachine/vendor.dm
@@ -74,6 +74,9 @@
 	if(istype(P, /obj/item/roguecoin/inqcoin))
 		return
 	if(istype(P, /obj/item/roguecoin))
+		if(!locked)
+			to_chat(user, span_warning("The Peddler is closed for purchases."))
+			return
 		budget += P.get_real_price()
 		qdel(P)
 		update_icon()
